refactor: improve some imports and names
